Fix numeric literal
