+++ crates/oxc_prettier/src/format/js.rs
@@ -906,64 +906,7 @@ impl<'a> Format<'a> for NullLiteral {
 
 impl<'a> Format<'a> for NumericLiteral<'a> {
     fn format(&self, p: &mut Prettier<'a>) -> Doc<'a> {
-        wrap!(p, self, NumericLiteral, {
-            // See https://github.com/prettier/prettier/blob/3.3.3/src/utils/print-number.js
-            // Perf: the regexes from prettier code above are ported to manual search for performance reasons.
-            let mut string = self.span.source_text(p.source_text).cow_to_ascii_lowercase();
-
-            // Remove unnecessary plus and zeroes from scientific notation.
-            if let Some((head, tail)) = string.split_once('e') {
-                let negative = if tail.starts_with('-') { "-" } else { "" };
-                let trimmed = tail.trim_start_matches(['+', '-']).trim_start_matches('0');
-                if trimmed.starts_with(|c: char| c.is_ascii_digit()) {
-                    string = Cow::Owned(std::format!("{head}e{negative}{trimmed}"));
-                }
-            }
-
-            // Remove unnecessary scientific notation (1e0).
-            if let Some((head, tail)) = string.split_once('e') {
-                if tail.trim_start_matches(['+', '-']).trim_start_matches('0').is_empty() {
-                    string = Cow::Owned(head.to_string());
-                }
-            }
-
-            // Make sure numbers always start with a digit.
-            if string.starts_with('.') {
-                string = Cow::Owned(std::format!("0{string}"));
-            }
-
-            // Remove extraneous trailing decimal zeroes.
-            if let Some((head, tail)) = string.split_once('.') {
-                if let Some((head_e, tail_e)) = tail.split_once('e') {
-                    if !head_e.is_empty() {
-                        let trimmed = head_e.trim_end_matches('0');
-                        if trimmed.is_empty() {
-                            string = Cow::Owned(std::format!("{head}.0e{tail_e}"));
-                        } else {
-                            string = Cow::Owned(std::format!("{head}.{trimmed}e{tail_e}"));
-                        }
-                    }
-                } else if !tail.is_empty() {
-                    let trimmed = tail.trim_end_matches('0');
-                    if trimmed.is_empty() {
-                        string = Cow::Owned(std::format!("{head}.0"));
-                    } else {
-                        string = Cow::Owned(std::format!("{head}.{trimmed}"));
-                    }
-                }
-            }
-
-            // Remove trailing dot.
-            if let Some((head, tail)) = string.split_once('.') {
-                if tail.is_empty() {
-                    string = Cow::Owned(head.to_string());
-                } else if tail.starts_with('e') {
-                    string = Cow::Owned(std::format!("{head}{tail}"));
-                }
-            }
-
-            dynamic_text!(p, &string)
-        })
+        literal::print_number(p, self.span.source_text(p.source_text))
     }
 }
